Add new separate delivery time entity.
