+++ custom_components/rohlikcz/translations/en.json
@@ -125,6 +125,9 @@
       },
       "delivery_info": {
         "name": "Delivery Information"
+      },
+      "delivery_time": {
+        "name": "Delivery Time"
       }
     },
     "todo": {
